moved function into d_graphs for better modularity
